add video and image publish simple

+++ TwitterApiSdk/Model/Respense/Upload/UploadFinalizeResponse.cs
@@ -9,7 +9,7 @@ namespace TwitterApiSdk.Model.Respense.Upload
         /// <summary>
         /// 
         /// </summary>
-        public int media_id { get; set; }
+        public long media_id { get; set; }
         /// <summary>
         /// 
         /// </summary>
